requirements.txt
livekit
livekit-agents
livekit-api
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-noise-cancellation
livekit-plugins-turn-detector
python-dotenv
pydantic
httpx>=0.27.0
